requirements.txt
requests
fastapi[standard]
uvicorn
YouTubeMusicAPI
ytmusicapi
pynput
pyannote.audio
torch[cpu]
rhasspy-client
paho-mqtt
pyowm
imap_tools